+++ packages/react/src/action-resolver-add-embed.ts
@@ -3,7 +3,7 @@ import {
   AddEmbedActionResolverEvents,
 } from "@mod-protocol/core";
 
-// This resolver is called when the miniapp calls the "ADDEMBED" action type.
+// This resolver is called when the mod calls the "ADDEMBED" action type.
 //
 // The action expected from the user is to update the current input field
 // (or equivalent) to embed the file specified in the init.
@@ -13,6 +13,6 @@ export default function actionResolverAddEmbed(
 ) {
   // eslint-disable-next-line no-console
   console.warn(
-    "Please implement 'AddEmbedActionResolver' and configure the MiniApp to use it"
+    "Please implement 'AddEmbedActionResolver' and configure the Mod to use it"
   );
 }