+++ packages/react/src/action-resolver-set-input.ts
@@ -3,7 +3,7 @@ import {
   SetInputActionResolverEvents,
 } from "@mod-protocol/core";
 
-// This resolver is called when the miniapp calls the "SETINPUT" action type.
+// This resolver is called when the mod calls the "SETINPUT" action type.
 //
 // The action expected from the user is to update the current input field with
 // the result of this action.
@@ -13,6 +13,6 @@ export default function actionResolverSetInput(
 ) {
   // eslint-disable-next-line no-console
   console.warn(
-    "Please implement 'SetInputActionResolver' and configure the MiniApp to use it"
+    "Please implement 'SetInputActionResolver' and configure the Mod to use it"
   );
 }